refactor: 포트폴리오 컴포넌트 세분화 진행
