--- src/components/portfolio/PortfolioPost.tsx
@@ -1,243 +1,0 @@
-import React, { forwardRef, useEffect, useRef, useState } from "react";
-import styled from "styled-components";
-import usePostBookmark from "hooks/portfolio/usePostPortfolio";
-import useGetMarkets from "hooks/useGetMarkets";
-
-type SearchMarket = { market: string; korean_name: string; english_name: string };
-
-interface ISearch {
-  market: string | undefined;
-  koreanName: string;
-}
-
-interface IPortfolio {
-  transaction: string;
-  price: number;
-  qty: number;
-  date: string;
-}
-
-const PortfolioPost = ({ setHeight }: { setHeight: React.Dispatch<React.SetStateAction<number>> }) => {
-  const { data: market } = useGetMarkets();
-  const { mutate: postPortfolio } = usePostBookmark();
-  const [search, setSearch] = useState<ISearch>({ market: "", koreanName: "" });
-  const [portforlio, setPortfolio] = useState<IPortfolio>({ transaction: "", price: 0, qty: 0, date: "" });
-
-  const searchValue = useRef<HTMLInputElement>(null);
-  const portRef = useRef<HTMLFormElement>(null);
-
-  const searchCoin = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const findMarket = market?.find((ele: any) => ele.korean_name === e.target.value);
-    setSearch({ ...search, koreanName: e.target.value, market: findMarket?.market });
-  };
-
-  const onChnagePortfolio = (e: React.ChangeEvent<HTMLInputElement> & React.MouseEvent<HTMLButtonElement>) => {
-    const { name, value } = e.target;
-    setPortfolio({ ...portforlio, [name]: value });
-  };
-
-  const postPortfolioHandelr = async (portforlio: IPortfolio) => {
-    const history = { ...portforlio, ...search };
-
-    if (!localStorage.getItem("accessToken")) {
-      alert("포트폴리오 기능은 로그인 후 이용할 수 있습니다.");
-      throw new Error("포트폴리오 기능은 로그인 후 이용할 수 있습니다.");
-    } else if (!history.market) {
-      alert("코인이름을 입력해주세요.");
-      throw new Error("코인이름을 입력해주세요.");
-    } else if (!history.qty) {
-      alert("수량을 입력해주세요.");
-      throw new Error("수량을 입력해주세요.");
-    } else if (!history.price) {
-      alert("금액을 입력해주세요.");
-      throw new Error("금액을 입력해주세요.");
-    } else if (!history.date) {
-      alert("날짜를 선택해주세요.");
-      throw new Error("날짜를 선택해 주세요.");
-    } else if (!history.transaction) {
-      alert("매수/매도를 선택해주세요.");
-      throw new Error("매수/매도를 선택해 주세요.");
-    }
-
-    postPortfolio(history);
-  };
-
-  const onSubmitPortfolio = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    postPortfolioHandelr(portforlio);
-  };
-
-  const searchMarket =
-    search.koreanName.length > 0 && market.filter((ele: any) => ele.korean_name.includes(search.koreanName));
-
-  useEffect(() => {
-    setHeight(portRef?.current?.clientHeight || 0);
-  }, [search]);
-
-  return (
-    <PostPortfolioContainer onSubmit={onSubmitPortfolio} ref={portRef}>
-      <InputContainer>
-        <div>
-          <UserInput
-            onChange={(e) => {
-              searchCoin(e);
-            }}
-            ref={searchValue}
-            placeholder="종목"
-            value={searchValue.current?.value || ""}
-            type="text"
-            name="market"
-          />
-          <UserInput placeholder="수량" type="number" name="qty" onChange={onChnagePortfolio} />
-        </div>
-        <div>
-          <UserInput placeholder="금액" type="number" step="0.000001" name="price" onChange={onChnagePortfolio} />
-          <UserInput placeholder="날짜" type="date" step="0.000001" name="date" onChange={onChnagePortfolio} />
-        </div>
-      </InputContainer>
-      <CoinSelect>
-        {searchMarket &&
-          searchMarket.map((ele: SearchMarket) => {
-            return (
-              <span
-                key={ele.market}
-                onClick={() => {
-                  searchValue?.current && (searchValue.current.value = ele.korean_name);
-                  setSearch({ market: ele.market, koreanName: ele.korean_name });
-                }}
-              >
-                {ele.korean_name}
-              </span>
-            );
-          })}
-      </CoinSelect>
-      <BtnContainer>
-        <div>
-          <BuyBtn
-            type="button"
-            name="transaction"
-            value="buy"
-            onClick={onChnagePortfolio}
-            transaction={portforlio.transaction}
-          >
-            매수
-          </BuyBtn>
-          <SellBtn
-            type="button"
-            name="transaction"
-            value="sell"
-            onClick={onChnagePortfolio}
-            transaction={portforlio.transaction}
-          >
-            매도
-          </SellBtn>
-        </div>
-        <Btn type="submit">포트폴리오 추가</Btn>
-      </BtnContainer>
-    </PostPortfolioContainer>
-  );
-};
-
-export default forwardRef(PortfolioPost);
-
-const PostPortfolioContainer = styled.form`
-  display: flex;
-  flex-direction: column;
-  min-height: 12%;
-  border-radius: 1rem;
-  background-color: #fff;
-  box-shadow: 0.25rem 0.25rem 0.5rem rgb(0 0 0 / 12%);
-  margin-bottom: 1rem;
-  padding: 2rem 1rem;
-`;
-
-const InputContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-
-  & div {
-    width: 49.5%;
-    display: flex;
-    justify-content: space-between;
-  }
-`;
-
-const UserInput = styled.input`
-  border-radius: 5px;
-  border: 1.5px solid #e5e7eb;
-  padding: 2%;
-  margin-bottom: 5px;
-  color: #7e7e7e;
-  width: 45%;
-
-  :focus {
-    outline: none;
-  }
-`;
-
-const CoinSelect = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  margin-bottom: 5px;
-
-  & span {
-    margin-right: 5px;
-    margin-bottom: 5px;
-    border: 1.5px solid #e5e7eb;
-    padding: 0.75rem 1rem;
-    font-size: 12px;
-    color: #808080;
-
-    :hover {
-      cursor: pointer;
-      border: 1.5px solid #3d6bfb;
-      color: #333;
-    }
-  }
-`;
-const BtnContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  & div {
-    display: flex;
-    justify-content: space-between;
-    width: 49.5%;
-
-    & button {
-      width: 49%;
-    }
-  }
-`;
-
-const Btn = styled.button`
-  width: 49.5%;
-  height: 3rem;
-  background-color: #1261c4;
-  border: none;
-  padding: 0;
-  border-radius: 5px;
-  margin-bottom: 5px;
-  color: #fff;
-  font-weight: 700;
-  font-size: 0.85rem;
-
-  :hover {
-    cursor: pointer;
-  }
-`;
-
-const BuyBtn = styled(Btn)<{ transaction: string }>`
-  border: 1px solid #f14f4f;
-
-  color: ${({ transaction }) => (transaction === "buy" ? "#fff" : "#f14f4f")};
-  background-color: ${({ transaction }) => (transaction === "buy" ? "#f14f4f" : "#fff")};
-`;
-
-const SellBtn = styled(Btn)<{ transaction: string }>`
-  border: 1px solid #3d6bfb;
-  color: ${({ transaction }) => (transaction === "sell" ? "#fff" : "#3d6bfb;")};
-  background-color: ${({ transaction }) => (transaction === "sell" ? "#3d6bfb" : "#fff")};
-`;
