+++ src/layout/Portfolio.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import PortfolioPost from "../components/portfolio/PortfolioPost";
+import PortfolioPost from "components/portfolio/postPortfolio/PortfolioPost";
 import styled from "styled-components";
 import PortfolioList from "../components/portfolio/PortfolioList";
